# apps/tenancy/tenancy/v1/tenancy.openapi.yaml
openapi: 3.1.0
info:
  title: Tenancy Service
  description: The Partition Service provides multi-tenancy and data isolation capabilities with hierarchical partitions, role-based access control (RBAC), and custom UI pages. Tenants represent top-level organizational units containing hierarchical partitions for departments, projects, or teams. The service supports fine-grained access control through partition roles and access grants, enabling secure multi-tenant applications with customizable user experiences.
  contact:
    name: Ant Investor Ltd
    url: https://github.com/antinvestor/service-authentication
    email: info@antinvestor.com
  license:
    name: Apache License
    url: https://github.com/antinvestor/apis/blob/master/LICENSE
  version: v1.0.0
paths:
...
  /tenancy.v1.TenancyService/GetClient:
    get:
      tags:
        - Clients
        - tenancy.v1.TenancyService
      summary: Get OAuth2 client
      description: Retrieves an OAuth2 client by its record ID or OAuth2 client_id.
      operationId: getClient
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
        - name: message
          in: query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenancy.v1.GetClientRequest'
        - name: encoding
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/encoding'
        - name: base64
          in: query
          schema:
            $ref: '#/components/schemas/base64'
        - name: compression
          in: query
          schema:
            $ref: '#/components/schemas/compression'
        - name: connect
          in: query
          schema:
            $ref: '#/components/schemas/connect'
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenancy.v1.GetClientResponse'
    post:
      tags:
        - Clients
        - tenancy.v1.TenancyService
      summary: Get OAuth2 client
      description: Retrieves an OAuth2 client by its record ID or OAuth2 client_id.
      operationId: getClient
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tenancy.v1.GetClientRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenancy.v1.GetClientResponse'
...
  /tenancy.v1.TenancyService/GetPartition:
    get:
      tags:
        - Partitions
        - tenancy.v1.TenancyService
      summary: Get partition
      description: Retrieves a partition by its unique identifier or domain, including all partition metadata and properties.
      operationId: getPartition
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
        - name: message
          in: query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenancy.v1.GetPartitionRequest'
        - name: encoding
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/encoding'
        - name: base64
          in: query
          schema:
            $ref: '#/components/schemas/base64'
        - name: compression
          in: query
          schema:
            $ref: '#/components/schemas/compression'
        - name: connect
          in: query
          schema:
            $ref: '#/components/schemas/connect'
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenancy.v1.GetPartitionResponse'
    post:
      tags:
        - Partitions
        - tenancy.v1.TenancyService
      summary: Get partition
      description: Retrieves a partition by its unique identifier or domain, including all partition metadata and properties.
      operationId: getPartition
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tenancy.v1.GetPartitionRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenancy.v1.GetPartitionResponse'
...
components:
  schemas:
    base64:
      type: boolean
      title: base64
      description: Specifies if the message query param is base64 encoded, which may be required for binary data
...
    compression:
      title: compression
      enum:
        - identity
        - gzip
        - br
      description: Which compression algorithm to use for this request
    connect:
      title: connect
      enum:
        - v1
      description: Define the version of the Connect protocol
    connect-protocol-version:
      type: number
      title: Connect-Protocol-Version
      enum:
        - 1
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
            - not_found
          enum:
            - canceled
            - unknown
            - invalid_argument
            - deadline_exceeded
            - not_found
            - already_exists
            - permission_denied
            - resource_exhausted
            - failed_precondition
            - aborted
            - out_of_range
            - unimplemented
            - internal
            - unavailable
            - data_loss
            - unauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/connect.error_details.Any'
          description: A list of messages that carry the error details. There is no limit on the number of messages.
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
    connect.error_details.Any:
      type: object
      properties:
        type:
          type: string
          description: 'A URL that acts as a globally unique identifier for the type of the serialized message. For example: `type.googleapis.com/google.rpc.ErrorInfo`. This is used to determine the schema of the data in the `value` field and is the discriminator for the `debug` field.'
        value:
          type: string
          format: binary
          description: The Protobuf message, serialized as bytes and base64-encoded. The specific message type is identified by the `type` field.
        debug:
          oneOf:
            - type: object
              title: Any
              additionalProperties: true
              description: Detailed error information.
          discriminator:
            propertyName: type
          title: Debug
          description: Deserialized error detail payload. The 'type' field indicates the schema. This field is for easier debugging and should not be relied upon for application logic.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message, with an additional debug field for ConnectRPC error details.
    encoding:
      title: encoding
      enum:
        - proto
        - json
      description: Define which encoding or 'Message-Codec' to use
...
    google.protobuf.Struct:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/google.protobuf.Value'
      description: |-
        `Struct` represents a structured data value, consisting of fields
         which map to dynamically typed values. In some languages, `Struct`
         might be supported by a native representation. For example, in
         scripting languages like JS a struct is represented as an
         object. The details of that representation are described together
         with the proto support for the language.

         The JSON representation for `Struct` is JSON object.
...
    tenancy.v1.ClientObject:
      type: object
      allOf:
        - properties:
            id:
              type: string
              title: id
              maxLength: 40
              minLength: 3
              pattern: '[0-9a-z_-]{3,40}'
              description: Unique client record ID
            name:
              type: string
              title: name
              maxLength: 100
              minLength: 3
              description: Display name
            clientId:
              type: string
              title: client_id
              maxLength: 100
              minLength: 3
              description: OAuth2 client_id (= Hydra client identifier)
            type:
              type: string
              title: type
              enum:
                - public
                - confidential
                - internal
                - external
              description: 'Client type: public, confidential, internal, external'
            grantTypes:
              type: array
              items:
                type: string
              title: grant_types
              description: OAuth2 grant types (authorization_code, refresh_token, client_credentials)
            responseTypes:
              type: array
              items:
                type: string
              title: response_types
              description: OAuth2 response types (code, token)
            redirectUris:
              type: array
              items:
                type: string
              title: redirect_uris
              description: Allowed redirect URIs
            scopes:
              type: string
              title: scopes
              description: Space-separated OAuth2 scopes
            audiences:
              type: array
              items:
                type: string
              title: audiences
              description: Service namespaces the client can access
            roles:
              type: array
              items:
                type: string
              title: roles
              description: Permission role names for SA tokens
            properties:
              title: properties
              description: Additional metadata
              $ref: '#/components/schemas/google.protobuf.Struct'
            state:
              title: state
              description: Client state
              $ref: '#/components/schemas/common.v1.STATE'
            createdAt:
              title: created_at
              description: Creation timestamp
              $ref: '#/components/schemas/google.protobuf.Timestamp'
        - oneOf:
            - type: object
              properties:
                partition:
                  title: partition
                  description: Partition that owns this client (user auth flows)
                  $ref: '#/components/schemas/tenancy.v1.PartitionObject'
              title: partition
              required:
                - partition
            - type: object
              properties:
                serviceAccount:
                  title: service_account
                  description: Service account that owns this client (client_credentials flows)
                  $ref: '#/components/schemas/tenancy.v1.ServiceAccountObject'
              title: service_account
              required:
                - serviceAccount
      title: ClientObject
      additionalProperties: false
      description: |-
        ClientObject represents an OAuth2 client configuration.
         A client can be owned by either a partition (for user auth flows like PKCE)
         or a service account (for client_credentials machine-to-machine flows).
...
    tenancy.v1.GetClientRequest:
      type: object
      properties:
        id:
          type: string
          title: id
          maxLength: 40
          minLength: 3
          pattern: '[0-9a-z_-]{3,40}'
          description: Client record ID
        clientId:
          type: string
          title: client_id
          maxLength: 100
          minLength: 3
          description: OAuth2 client_id
      title: GetClientRequest
      additionalProperties: false
    tenancy.v1.GetClientResponse:
      type: object
      properties:
        data:
          title: data
          $ref: '#/components/schemas/tenancy.v1.ClientObject'
      title: GetClientResponse
      additionalProperties: false
...
    tenancy.v1.GetPartitionRequest:
      type: object
      properties:
        id:
          type: string
          title: id
          maxLength: 40
          minLength: 3
          pattern: '[0-9a-z_-]{3,40}'
          description: Partition ID
        domain:
          type: string
          title: domain
          maxLength: 255
          description: Partition domain (unique)
      title: GetPartitionRequest
      additionalProperties: false
    tenancy.v1.GetPartitionResponse:
      type: object
      properties:
        data:
          title: data
          $ref: '#/components/schemas/tenancy.v1.PartitionObject'
      title: GetPartitionResponse
      additionalProperties: false
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security: []
...